timeline times out to go to welcome page
